refactor getDisplayPages to use util methods
