+++ src/components/pagination/pagination.component.js
@@ -1,8 +1,8 @@
 // https://reactjs.org/docs/forms.html
-
 import React, { Component, Fragment } from 'react'
+import { endsWith, equal, last } from '@code.gov/cautious'
 
-import { endsWith, equal, last, range } from '@code.gov/cautious'
+import { getPageInfo, getPaginationNavInfo } from '../../utils/pagination'
 
 export default class Pagination extends Component {
 
@@ -57,18 +57,11 @@ export default class Pagination extends Component {
   getDisplayPages() {
     const { count, pagesize } = this.props
     const page = parseInt(this.props.page, 10)
-
-    const pagecount = Math.ceil(count / pagesize)
-    const pageIndexes = range(pagecount)
-      .map(n => n + 1) // convert from starting at 0 to 1
-
-    const pageCount = pageIndexes.length
+    const { pageCount, pageIndexes } = getPageInfo({ count, pagesize })
     let displayPages = []
 
     try {
-      const ultimate = last(pageIndexes)
-      const left = page - 1
-      const right = page + 1
+      const { ultimate, left, right } = getPaginationNavInfo({ page, pageIndexes })
 
       if (pageCount <= 7) {
         displayPages = pageIndexes
